+++ ApiKarbord/Models/Public/Web_IDocH.cs
@@ -173,7 +173,7 @@
 
         public double? ArzRate { get; set; }
 
-        public int? AccSerialNumber { get; set; }
+        public long? AccSerialNumber { get; set; }
 
         public int? AccDocNo { get; set; }
